% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{add_clustering_annotation}
\alias{add_clustering_annotation}
\title{Add Clustering Annotations}
\usage{
add_clustering_annotation(dataList, memberships, clusteringName)
}
\arguments{
\item{dataList}{A list of data sets to add the clustering annotation too}

\item{memberships}{A list of the cluster memberships of each cell, used to add the annotation}

\item{clusteringName}{The name of the clustering method used to obtain the clustering membership}
}
\value{
A list of the data sets with the clustering annotations added
}
\description{
This function will add clustering annotations (i.e memberships) to the metadata of the Seurat Object of the matching data.
Any clustering method can be used to obtain these memberships. The annotation name will include the clustering type
and the cells cluster ID.
}
