% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{add_duplicate_cells}
\alias{add_duplicate_cells}
\title{Add duplicated Cells}
\usage{
add_duplicate_cells(dataList)
}
\arguments{
\item{dataList}{a  list of data sets}
}
\value{
A data set list with a duplicated cell sample
}
\description{
This function takes a random cell sample and adds its duplicate to the data set. This function is a test of algorithmic stability
}
