% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{add_zero_variance_gene_counts}
\alias{add_zero_variance_gene_counts}
\title{Add Zero Variance Gene Counts}
\usage{
add_zero_variance_gene_counts(dataList)
}
\arguments{
\item{dataList}{a  list of data sets}
}
\value{
A data set list with added zero variance gene counts
}
\description{
This function adds a set of zero variance gene counts to the sample. This function is a test of algorithmic stability
}
