% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{annotate_seurat_object}
\alias{annotate_seurat_object}
\title{Annotate Seurat Object}
\usage{
annotate_seurat_object(data)
}
\arguments{
\item{data}{A Seurat object that you want annotated}
}
\value{
A annotated Seurat Object
}
\description{
This function annotates a created Seurat object. It edits the chunks and meta.data in
order to annotate the object for use in down stream analysis
}
