% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{complete_kmeans}
\alias{complete_kmeans}
\title{Completes the KMeans Process}
\usage{
complete_kmeans(datasets, seed = 1, path = ".")
}
\arguments{
\item{datasets}{The names of the datasets for the completion of the workflow}

\item{seed}{The seed of randomization for reproducibility, defaults to 1}

\item{path}{The file path for saving the results of the workflow}
}
\description{
This function completes the entire kmeans workflow procedure for the log data transformation. This workflow handles data transformation,
dimensionality reduction, and clustering. The results will be stored in a csv file called kmeans_clusters.csv
}
