% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{complete_seurat_gficf}
\alias{complete_seurat_gficf}
\title{Completes the Seurat Process on GFICF data}
\usage{
complete_seurat_gficf(datasets, seed = 1, path = ".", reduction = "pca")
}
\arguments{
\item{datasets}{The names of the datasets for the completion of the workflow}

\item{seed}{The seed of randomization for reproducibility, defaults to 1}

\item{path}{The file path for saving the results of the workflow}
}
\description{
This function completes the entire Seurat workflow procedure for the GFICF data transformation. This workflow handles data transformation,
dimensionality reduction, and clustering. The results will be stored in a csv file called seurat_clusters_gficf.csv
}
