% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_internal_cluster_validation_report}
\alias{create_internal_cluster_validation_report}
\title{Create Internal Cluster Validation Report}
\usage{
create_internal_cluster_validation_report(dataList, method = "kmeans")
}
\arguments{
\item{dataList}{A list of datasets with clustering completed}

\item{method}{The type of clustering method to validate, defaults to kmeans}
}
\value{
A Internal Cluster Validation Report table with the Number of clusters, Average Silhouette Width,
and Dunn Index as measures of validation
}
\description{
This function creates an internal validation report for clustering methods. This report includes the
Average Silhouette Width, as well as the Dunn Index as measures of Internal Validation. It also includes other useful information
such as the Number of Clusters and the Number of Singletons.
}
