% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_metamorphic_testing_report}
\alias{create_metamorphic_testing_report}
\title{Create Metamorphic Testing Report}
\usage{
create_metamorphic_testing_report(reportList)
}
\arguments{
\item{reportList}{a list of internal clustering validation reports contain information from metamorphic testing}
}
\value{
A report of the percent of reductions that passed the metamorphic test for each test and each dataset
}
\description{
This function creates a metamorphic testing report for clustering methods. This report includes which
reductions of the dataset passed the metamorphic test.
}
