% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{create_seurat_plot}
\alias{create_seurat_plot}
\title{Visualization method for the Seurat Clustering}
\usage{
create_seurat_plot(data, reduction_choosen = "pca")
}
\arguments{
\item{data}{An individual data object that you want to visualize}

\item{reduction_choosen}{The dimensionality reduction method choosen}
}
\description{
Produces a plot of the Seurat clustering, based on the dimensionality reduction method chosen. The plot is a Seurat DimPlot
}
