% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_download}
\alias{data_download}
\title{Downloads the data}
\usage{
data_download(path = ".")
}
\arguments{
\item{path}{The path of the download location to use}
}
\description{
This function downloads the data that is used in the package for use in examples, and automatically loads the data for
usage. Run this once when the data is not download, if the data is downloaded use the data_load function
}
