% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{describe_datasets}
\alias{describe_datasets}
\title{Describe Datasets}
\usage{
describe_datasets(dataList)
}
\arguments{
\item{dataList}{A list of names of data sets that you would like to view}
}
\description{
This function displays a short summary of each data set in the data list provided. Can be used to gain quick insights into the
information that is stored in the data
}
