% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_data}
\alias{download_data}
\title{Downloads the data}
\usage{
download_data(path = "./")
}
\arguments{
\item{path}{The path of the download location to use}
}
\value{
A object with the names of the data sets that were downloaded and loaded
}
\description{
This function downloads the data that is used in the package for use in examples, and automatically loads the data for
usage. Run this once when the data is not downloaded, if the data is downloaded use the load_data function
}
