% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{duplicate_datasets}
\alias{duplicate_datasets}
\title{Duplicate Datasets}
\usage{
duplicate_datasets(dataList, duplicates = 2)
}
\arguments{
\item{dataList}{A list of data set names that should be duplicated}

\item{duplicates}{The number of duplicates that are desired, defaults to 2}
}
\value{
a data list of the duplicated datasets
}
\description{
This function duplicates a data set a variable number of times. This function is to be used in
research situations.
}
