% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extract_datasets}
\alias{extract_datasets}
\title{Extract Datasets}
\usage{
extract_datasets(names)
}
\arguments{
\item{names}{A list of names representing the data sets you would like to extract}
}
\value{
A data list of the matrices of the data sets that were extracted
}
\description{
This function extracts a set of Data sets for further use in the workflows. Each data set that is extracted will be a dgCMatrix
with rows representing genes and cells representing columns
}
