% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filter_data}
\alias{filter_data}
\title{Filter the a Data List}
\usage{
filter_data(dataList)
}
\arguments{
\item{dataList}{A list of data, with Genes in Rows and Cells in Columns}
}
\value{
A data list of data that falls within the specified data limitations which acts as preprocesses
}
\description{
This function "filters" a data list by restricting a input list to a range between minimum and maximum of the Total MRNA counts
Total Gene Counts that are present in the list. This creates a filtered set of genes to use in experiments, removing extreme
outliers and making sure that all cells and genes are potentially informative.
}
