% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{find_num_cluster}
\alias{find_num_cluster}
\title{Find the number of clusters}
\usage{
find_num_cluster(dataList, clusteringType = "kmeans", reductionType = "pca")
}
\arguments{
\item{dataList}{A data list of data sets}

\item{clusteringType}{the type of clustering you would like to find the number of}

\item{reductionChoosen}{The dimensionality reduction type that is wanted}
}
\value{
A results list with the number of clusters
}
\description{
This functions finds the number of clusters in the Seurat Clustering method, or through the kmeans clustering
method.
}
