% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{flip_gene_counts}
\alias{flip_gene_counts}
\title{Flips the Gene Counts}
\usage{
flip_gene_counts(dataList)
}
\arguments{
\item{dataList}{a  list of data sets}
}
\value{
A data set list with flipped gene counts
}
\description{
This function multiplies the gene count values by -1, negating their values. This function is a test of algorithmic stability of a pipeline or method.
}
