% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_num_clusters}
\alias{get_num_clusters}
\title{Get Number of CLusters}
\usage{
get_num_clusters(dataList, reductionType = "pca", method = "kmeans")
}
\arguments{
\item{dataList}{A list of datasets with clustering completed}

\item{reductionType}{The type of reduction to use}

\item{method}{The type of clustering method to validate, defaults to kmeans}
}
\value{
A list with the number of clusters
}
\description{
This function finds the number of clusters per data set.
}
