% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_number_clusters}
\alias{get_number_clusters}
\title{Get the number of clusters}
\usage{
get_number_clusters(dataList, reductionType = "pca", method = "kmeans")
}
\arguments{
\item{dataList}{A data list of data sets}

\item{reductionType}{The type of dimensionly reduced data that should be used for clustering,
defaults to PCA}

\item{method}{The type of clustering, defaults to kmeans}
}
\value{
A list with the number of clusters
}
\description{
This functions finds the number of clusters in the Seurat Clustering method, or through the kmeans clustering
method.
}
