% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_number_singletons}
\alias{get_number_singletons}
\title{Get the number of Singletons}
\usage{
get_number_singletons(dataList, reductionType = "pca", method = "kmeans")
}
\arguments{
\item{dataList}{A data list of data sets}

\item{reductionType}{The type of dimensionly reduced data that should be used for clustering,
defaults to PCA}

\item{method}{The type of clustering, defaults to kmeans}
}
\value{
A list with the number of Singleton Clusters
}
\description{
This functions finds the number of Singletons Clusters in the Seurat Clustering method, or through the kmeans clustering
method.
}
