% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_scenario}
\alias{get_scenario}
\title{Get Scenario}
\usage{
get_scenario(scenario_id)
}
\arguments{
\item{scenario_id}{The scenario_id that you would like to identify}
}
\description{
This function gets the Scenario used in the creation of the data and the sequencing
technique used in collection. Only to be used internally.
}
