% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{gficf_workflow}
\alias{gficf_workflow}
\title{GFICF Workflow}
\usage{
gficf_workflow(idx, seed = 1)
}
\arguments{
\item{idx}{The names of the datasets}

\item{seed}{The random seed that you would like to use on the dataset}
}
\value{
a result list with the dataset ID and the cluster it belongs to
}
\description{
This workflow is designed to apply the GFICF data transformation on the datasets
and test it's ability to handle pca, umap, and clustering.
}
