% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merges Data sets}
\usage{
merge_datasets(data.list, intersect = TRUE)
}
\arguments{
\item{data.list}{A list of data set to merge}

\item{intersect}{A Boolean parameter, that decides if the output should be a sparse matrix or a matrix}
}
\value{
A data list of the merged data
}
\description{
This function merges a series of data lists into one data matrix, based on the intersection of the individuals. If there is intersection
the matrix is merged as sparse to handle zeros, otherwise it is merged as a matrix. This function assumes that the data is
in the format genes in rows and cells in columns.
}
