% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{merge_sparse}
\alias{merge_sparse}
\title{Merge the Sparse Matrices}
\usage{
merge_sparse(...)
}
\description{
This function merges the contents of two lists into a sparse matrix. It assumes that the data sets
are of the same format, with genes stored in rows and cells stored in columns. The lists are merged base on the
union of the row names and the union of the column names
}
\details{
TODO: FIX arguement passing
}
