% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{permute_dataset_order}
\alias{permute_dataset_order}
\title{Permute the Order fo the data sets}
\usage{
permute_dataset_order(dataList)
}
\arguments{
\item{dataList}{a  list of data sets}
}
\value{
A permuted data set list
}
\description{
This function permutes the order of the data set
}
