% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{permute_rows}
\alias{permute_rows}
\title{Permute the Rows of a Data List}
\usage{
permute_rows(dataList)
}
\arguments{
\item{dataList}{A list of data}
}
\value{
The data list with the rows permuted
}
\description{
This function is designed for experimental analysis of algorithmic stability. It permutes the rows of a data list
so that further experimentation can be done with the data list to check the  algorithmic stability of a pipeline or
method
}
