% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{perturb_datasets}
\alias{perturb_datasets}
\title{Preturb Data sets}
\usage{
perturb_datasets(dataList, perturbationType = 1)
}
\arguments{
\item{dataList}{A list of data set that should be perturbed}

\item{perturbationType}{The number of the perturbation that is too be applied}
}
\value{
a data list of the perturbed datasets
}
\description{
This function applies a specific perturbation to the data-list based on the perturbation type. For a description of the
perturbations please see the individual functions
}
