% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess}
\alias{preprocess}
\title{Preprocesses the a Data List}
\usage{
preprocess(data.list)
}
\arguments{
\item{data.list}{A list of data, with Genes in Rows and Cells in Columns}
}
\value{
A data list of data that falls within the specified data limitations which acts as preprocesses
}
\description{
This function "preprocesses" a data list by restricting a input list to a range between minimum and maximum of the Total MRNA counts
Total Gene Counts that are present in the list. This creates a filtered set of genes to use in experiments, removing extreme
outliers. These bounds are similar to the ?monocle? package. CITE PAPER
}
