% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{run_ari}
\alias{run_ari}
\title{Run ARI}
\usage{
run_ari(dataList, groundTruths, reductionType = "pca", method = "kmeans")
}
\arguments{
\item{dataList}{A list of data sets with cluster membership information}

\item{groundTruths}{The true cell cluster memberships}

\item{reductionType}{The dimensionality reduction type of the data to find the calculated scores,
defaults to PCA}

\item{method}{The clustering method used, defaults to Kmeans}
}
\value{
A table of the ARI Scores, with the data set name and the score
}
\description{
Computes the Adjusted Rand Index of the cluster memberships. Uses the aricode implementation of ARI
}
