% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_cca}
\alias{run_cca}
\title{Run CCA}
\usage{
run_cca(dataList)
}
\arguments{
\item{dataList}{A list of data sets to integrate using the cca protocol}
}
\value{
A data list of the combined data from the cca protocol
}
\description{
This function runs the data integration protocol detailed in the Seurat "Introduction to scRNA-seq
integration" found at https://satijalab.org/seurat/articles/integration_introduction.html. This function completes the cca integration process using the first data set as a reference as described at https://satijalab.org/seurat/articles/integration_large_datasets.html
Use run_cca_integration_workflow() for the entire workflow process
}
