% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{run_cluster}
\alias{run_cluster}
\title{Complete Clustering Steps}
\usage{
run_cluster(data.list, reduction_choosen = "pca")
}
\arguments{
\item{data.list}{A data list of data sets}
}
\value{
A data list with clustering completed completed on the features
}
\description{
This functions runs the Seurat FindNeighbors and FindCLusters function on a list of data sets.
This functions assumes that genes are in rows and
cells are in columns. The FindNeighbors reduction is set to "pca" by default and uses the first 30 dimensions
The FindClusters resolution is set to 0.5
}
