% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_clustering_workflow}
\alias{run_clustering_workflow}
\title{Completes the clustering workflow}
\usage{
run_clustering_workflow(
  dataList,
  method = "kmeans",
  transformationType = "log",
  seed = 1,
  numberClusters = 10
)
}
\arguments{
\item{dataList}{A list of the data sets}

\item{method}{The method of clustering to use, defaults to Kmeans}

\item{transformationType}{The data transformation to be used, defaults to the log transformation}

\item{seed}{The seed to be set for reproducibility, defaults to 1}

\item{numberClusters}{The number of clusters to use for Kmeans clustering, defaults to 10}
}
\value{
the Data list with all components of the workflow completed
}
\description{
This function completes the entire clustering workflow, on different data transformation types,
and with different clustering methods. The log data transformation pipeline includes filter_dataing, annotating,
applying the log transformation, selecting the Highly Variable Genes, and finally scaling the data before
dimensionality reduction, and clustering. The TFIDF data transformation completes the TFIDF data transformation,
annotates the data, and selects the Highly Variable Genes before applying dimensionality reduction and finally
clustering. The clustering parameters are set to default, except for the number of clusters in
kmeans clustering, which is set to 10
}
