% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{run_dunn}
\alias{run_dunn}
\title{Run Dunn}
\usage{
run_dunn(dataList, reduction_choosen, method = "kmeans")
}
\arguments{
\item{dataList}{a data list of the data with clustering assignments}

\item{reduction_choosen}{The dimensionality reduction type of the data to find the calculated scores}

\item{method}{The clustering method used}
}
\description{
This function calculates the dunn index of the clustering assignments made on the datasets provided
}
