% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow2.R
\name{run_duplicate_integrations}
\alias{run_duplicate_integrations}
\title{Run Integration Duplicated (Need a better name)}
\usage{
run_duplicate_integrations(datasets, dup)
}
\arguments{
\item{datasets}{A list of data set names that you would like to integrate with
the 4 pipelines}

\item{dup}{The number of duplications that would like to execute}
}
\description{
This functions completes an integration pipeline on a dupilcated set of the original data, using
4 integration techniques. It implements the Harmony, FastMNN, CCA, and scTranform integration techniques.
These are implemented via the Seurat and Harmony packages. See ?run_harmony, ?run_cca, ?run_fastmnn, and ?run_sctransform
for more information on these pipeline structures. Each of these integration will write the integrated data set to a file
in the current directory
}
