% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_fastmnn}
\alias{run_fastmnn}
\title{Run fastmnn}
\usage{
run_fastmnn(dataList, batchName = "ID", seed = 1)
}
\arguments{
\item{dataList}{A data list of data sets to integrate using the fastmnn protocol}
}
\value{
data.combined A data list of the combined data from the fastmnn protocol
}
\description{
This function runs the data integration protocol detailed in the Seurat ## FIND VING. AND LINK . This function completes the fastmnn process on
the datasets. Use run_fastmnn_workflow() for the entire workflow process
}
