% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{run_gficf}
\alias{run_gficf}
\title{Runs the GFICF Transformations}
\usage{
run_gficf(dataList)
}
\arguments{
\item{dataList}{A data list of Data sets in the Gene in Row and Cell in Columns format}
}
\value{
A data list with GFICF scores of Each Gene Cell Relationship
}
\description{
This function completes the Gene Frequency Inverse Cell Frequency data transformation on a list of data. This is step is complete by
using the text2vec protocol for completing TF-IDF on the cell data, creating GF-ICF Scores
}
