% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_harmony}
\alias{run_harmony}
\title{Run Harmony}
\usage{
run_harmony(dataList, batchName = "ID")
}
\arguments{
\item{dataList}{A list of data sets to integrate using the harmony protocol}

\item{batchName}{The annotation name to group data sets by, defaults to "ID"}
}
\value{
A list of the integrated data from the harmony protocol
}
\description{
This function runs the data integration protocol found at https://github.com/satijalab/seurat-wrappers/blob/master/docs/harmony.md
This function completes the harmonization process on
the data sets. Use run_harmony_integration_workflow() for the entire workflow process
}
