% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_harmony_integration_workflow}
\alias{run_harmony_integration_workflow}
\title{Completes the Harmony Workflow}
\usage{
run_harmony_integration_workflow(
  dataList,
  method = "kmeans",
  numberClusters = 10,
  seed = 1
)
}
\arguments{
\item{dataList}{A list of data sets to merge}

\item{method}{The clustering method to use}

\item{numberClusters}{The number of clusters to use in the kmeans clustering}

\item{seed}{The seed of randomization for reproducibility, defaults to 1}
}
\value{
A list of datasets with the harmony workflow applied
}
\description{
This function completes the harmony workflow procedure for the log data transformation for merging datasets. This workflow handles data transformation, variable gene selection,
scaling, dimensionality reduction, integration, and clustering.
}
