% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_harmony_workflow}
\alias{run_harmony_workflow}
\title{Completes the Harmony Workflow on Log Data}
\usage{
run_harmony_workflow(datasets, seed = 1)
}
\arguments{
\item{datasets}{The names of the datasets for the completion of the workflow}

\item{seed}{The seed of randomization for reproducibility, defaults to 1}
}
\description{
This function completes the harmony workflow procedure for the log data transformation for merging datasets. This workflow handles data transformation,
dimensionality reduction, merging, and clustering.
}
