% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow2.R
\name{run_integration}
\alias{run_integration}
\title{Run Integration  (Need a better name)}
\usage{
run_integration(datasets)
}
\arguments{
\item{datasets}{A list of data set names that you would like to integrate with
the 4 pipelines}
}
\description{
This functions completes an integration pipeline, using
4 integration techniques. It implements the Harmony, FastMNN, CCA, and scTranform integration techniques.
These are implemented via the Seurat and Harmony packages. See ?run_harmony, ?run_cca, ?run_fastmnn, and ?run_sctransform
for more information on these pipeline structures. Each of these integration will write the integrated data set to a file
in the current directory
}
