% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{run_kmeans}
\alias{run_kmeans}
\title{Run Kmeans}
\usage{
run_kmeans(dataList, k = 10, reductionChoosen = "pca")
}
\arguments{
\item{dataList}{A list of datasets to apply the Kmeans algorithm too}

\item{k}{The number of cluster centers to use, defaults to 10}
}
\value{
The dataList with kmeans clusters stored in the meta data
}
\description{
This function runs Kmeans Clustering. It uses the simple Kmeans algorithm with
a max number of iterations set to be 100.
}
