% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{run_kmeans_clustering}
\alias{run_kmeans_clustering}
\title{Run Kmeans}
\usage{
run_kmeans_clustering(
  dataList,
  reductionType = "pca",
  numberClusters = 10,
  iterationsMax = 100,
  seed = 1
)
}
\arguments{
\item{dataList}{A list of data sets to be clustered by Kmeans}

\item{reductionType}{The type of dimensionly reduced data that should be used for clustering,
defaults to PCA}

\item{numberClusters}{The number of cluster centers to use, defaults to 10}

\item{iterationsMax}{The max number of iterations that should be used in the kmeans algorithm,
defaults to 100}
}
\value{
The dataList with kmeans clusters stored in the meta data
}
\description{
This function runs Kmeans Clustering. It uses the simple Kmeans algorithm with
a max number of iterations set to be 100 by default. The number of clusters can be set by the users and defaults
to 10 clusters. This is the R package stats implementation of Kmeans
}
