% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{run_log}
\alias{run_log}
\title{Runs the Log Transformation}
\usage{
run_log(data.list)
}
\arguments{
\item{data.list}{A data list of Data sets in the Gene in Row and Cell in Columns format}
}
\value{
A data list with Log scores of Each Gene Cell Relationship
}
\description{
This function completes the Log data transformation on a list of data. This step can be done to
create and use Log data for further downstream data analysis and computations. For more information on the Log calculation
please see https://cran.r-project.org/web/packages/Seurat/Seurat.pdf
}
