% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{run_metamorphic_test}
\alias{run_metamorphic_test}
\title{Run metamorphic tests}
\usage{
run_metamorphic_test(
  dataList,
  perturbationType = 1,
  method = "kmeans",
  transformationType = "log",
  seed = 1,
  numberClusters = 10
)
}
\arguments{
\item{dataList}{A list of data sets}

\item{method}{The clustering method to be used, defaults to Kmeans}

\item{transformationType}{The type of data transformation to apply, defaults to log}

\item{seed}{the seed to be set for reproducibility, defaults to 1}

\item{numberClusters}{The number of cluster centers to use, defaults to 10}
}
\value{
A metamorphic results table
}
\description{
This function runs the metamorphic test. It applies a perturbation and then clusters based on the method, transformation, and number of
clusters supplied.
}
