% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_seurat}
\alias{run_seurat}
\title{Runs the Seurat Pipeline}
\usage{
run_seurat(dataList, seed = 1)
}
\arguments{
\item{dataList}{A data list of Data sets in the Gene in Row and Cell in Columns format}
}
\description{
This function completes the Seurat Pipeline on a list of data. This process starts with finding the variable features (genes), then
scaling the data, running PCA, UMPA, and then find clusters.  This step can be done to test your benchamrk data sets
or for down stream analysis. For more information on the Seurat Pipeline
please see https://satijalab.org/seurat/index.html
}
