% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{run_seurat_clustering}
\alias{run_seurat_clustering}
\title{Seurat Clustering}
\usage{
run_seurat_clustering(
  dataList,
  reductionType = "pca",
  resolutionValue = 0.5,
  numberComponents = 10
)
}
\arguments{
\item{dataList}{A list of data sets to be clustered}

\item{reductionType}{The type of dimensionly reduced data that should be used for clustering, defaults to PCA}

\item{resolutionValue}{The resolution for the Seurat clustering method, defaults to 0.5}

\item{numberComponents}{the number of components to use, defaults to 10}
}
\value{
A data list with Seurat clustering completed
}
\description{
This functions runs the Seurat FindNeighbors and FindClusters function on a list of data sets.
This functions assumes that genes are in rows and
cells are in columns. The FindNeighbors reduction is set to "pca" by default and uses the first 10 dimensions
The FindClusters resolution is set to 0.5 by default, but is controlable by the user
}
