% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_seurat_columns}
\alias{run_seurat_columns}
\title{Run the Seurat Pipeline with column Permutations}
\usage{
run_seurat_columns()
}
\value{
a results matrix with ARI and idx number
}
\description{
This function runs the built Seurat pipeline with column permutations on a data list. The results are appened into
the rows of a new matrix, with the idx and the ARI recorded
}
