% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_seurat_rows}
\alias{run_seurat_rows}
\title{Run the Seurat Pipeline with row Permutations}
\usage{
run_seurat_rows()
}
\value{
a results matrix with ARI and idx number
}
\description{
This function runs the built Seurat pipeline with row permutations on a data list. The results are appended into
the rows of a new matrix, with the idx and the ARI recorded
}
