% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{run_tfidf}
\alias{run_tfidf}
\title{Runs the tfidf Transformations}
\usage{
run_tfidf(dataList)
}
\arguments{
\item{dataList}{A data list of Data sets in the Gene in Row and Cell in Columns format}
}
\value{
A data list with TFIDF scores of each Gene-Cell relationship
}
\description{
This function completes the Term Frequency Inverse Document Frequency data transformation on a list of data. This is step is completed by
using the text2vec protocol for TF-IDF on the cell data, creating TF-IDF Scores
}
