% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{scale_data}
\alias{scale_data}
\title{Scales the Data}
\usage{
scale_data(data.list)
}
\arguments{
\item{data.list}{A list of data sets that you would like to scale}
}
\value{
A data list of that includes the scaled data
}
\description{
This function uses the ScaleData of Seurat to scale the data set. This function
assumes that genes are in rows and cells are in columns.
}
