% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{test_gficf_workflow}
\alias{test_gficf_workflow}
\title{Test GFICF Workflow}
\usage{
test_gficf_workflow()
}
\arguments{
\item{idx}{The names of the datasets}
}
\description{
This function is designed to apply the GFICF data transformation workflow on the datasets
and test it's ability to handle pca, umap, and clustering.
}
