% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{test_log_workflow}
\alias{test_log_workflow}
\title{Test Log Workflow}
\usage{
test_log_workflow()
}
\arguments{
\item{datasets}{The names of the datasets}
}
\description{
This function is designed to apply the log data transformation workflow on the datasets
and test it's ability to handle pca, umap, and clustering.
}
