% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{view_data}
\alias{view_data}
\title{View Data}
\usage{
view_data(idx)
}
\arguments{
\item{dataList}{A list of data sets that you would like to view}
}
\description{
This function displays a short summary of each data set in the data list provided. Can be used to gain quick insights into the
information that is stored in the data
}
