% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{write_output}
\alias{write_output}
\title{Write Output}
\usage{
write_output(data, prefix)
}
\arguments{
\item{data}{The data the needs to be output}

\item{prefix}{The prefix of the integratio technique used}
}
\description{
This function writes the output of the data integration pipelines
to a file with the associated labels, and data information.
}
