% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{build_seurat_columns}
\alias{build_seurat_columns}
\title{Build Seurat Pipeline with the Columns Permuted}
\usage{
build_seurat_columns(idx, seed = 1)
}
\arguments{
\item{idx}{A data set with genes as rows and cells as columns}
}
\value{
Adjusted Rand Index
}
\description{
This functions builds the Seurat Pipeline, with experimental condition. The initial data set is appended with a the same data list,
with permuted columns. This allows the testing of the algorithmic stability of the Seurat Pipeline for the specified data list. The
Normalized information Distance is returned. If there is true algorithmic stability this score would be 1
}
