% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{duplicate_datasets}
\alias{duplicate_datasets}
\title{Duplicate Datasets}
\usage{
duplicate_datasets(dataList, duplicates = 2)
}
\arguments{
\item{names}{A list of dataset names that should be duplicated}

\item{ndups}{The number of duplicates that are desired}
}
\value{
a dataList of the duplicated datasets
}
\description{
This function duplicates a dataset a variable number of times. This function is to be used in
research situations.
}
