% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merge the Sparse data sets}
\usage{
merge_datasets(dataList)
}
\arguments{
\item{dataList}{A list of data that  you would like to merge}
}
\value{
A list of the merged data sets
}
\description{
This function merges the contents of two lists into a sparse matrix. It assumes that the data sets
are of the same format, with genes stored in rows and cells stored in columns. The lists are merged base on the
union of the row names and the union of the column names
}
