% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{permute_columns}
\alias{permute_columns}
\title{Permute the columns of a Data List}
\usage{
permute_columns(data.list)
}
\arguments{
\item{data.list}{A data list that you would like to permute the columns of}
}
\value{
The data list with the columns permuted
}
\description{
This function is designed for experimental analysis of algorithmic stability. It permutes the columns of a data list
so that further experimentation can be done wit the data list to check the  algorithmic stability of a pipeline or
method
}
