% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{permute_dataset_order}
\alias{permute_dataset_order}
\title{Permute the Order fo the data sets}
\usage{
permute_dataset_order(dataList)
}
\arguments{
\item{dataList}{a  list of data sets}
}
\value{
A permuted data set list
}
\description{
This function is designed for experimental analysis of algorithmic stability.
This function permutes the order of the data set so that further experimentation can be done with the data list to check the  algorithmic stability of a pipeline or
method
}
