% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Visualization method for the Clustering Methods}
\usage{
plot_clusters(data, reductionChoosen = "pca", method = "kmeans")
}
\arguments{
\item{data}{An individual data object that you want to visualize}

\item{reductionChoosen}{The dimensionality reduction method choosen}

\item{method}{The cluster method to use}
}
\description{
Produces a plot of the clustering, based on the dimensionality reduction method chosen. The plot is a Seurat DimPlot
}
