% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{run_dunn}
\alias{run_dunn}
\title{Run Dunn Index}
\usage{
run_dunn(dataList, reductionType = "pca", method = "kmeans")
}
\arguments{
\item{dataList}{a data list of the data with clustering assignments}

\item{reductionType}{The dimensionality reduction type of the data to find the calculated scores,
defaults to PCA}

\item{method}{The clustering method used, defaults to Kmeans
#' @return A table of the Dunn Index Scores, with the data set name and the score}
}
\description{
This function calculates the Dunn index of the clustering assignments made on the datasets provided. Uses the
clVaild implementation of the Dunn Index
}
