% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_fastmnn}
\alias{run_fastmnn}
\title{Run fastmnn}
\usage{
run_fastmnn(dataList, batchName = "ID")
}
\arguments{
\item{dataList}{A  list of data sets to integrate using the fastmnn protocol}

\item{batchName}{The annotation name to group data sets by, defaults to "ID"}
}
\value{
A data list of the combined data from the fastmnn protocol
}
\description{
This function runs the data integration protocol detailed in the Seurat Wrappers protocol https://github.com/satijalab/seurat-wrappers/blob/master/docs/fast_mnn.md
This function completes the fastmnn process on
the datasets. Use run_fastmnn_integration_workflow() for the entire workflow process
}
