% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{run_gficf}
\alias{run_gficf}
\title{Runs the GFICF Transformations}
\usage{
run_gficf(data.list)
}
\arguments{
\item{data.list}{A data list of Data sets in the Gene in Row and Cell in Columns format}
}
\value{
A data list with GFICF scores of Each Gene Cell Relationship
}
\description{
This function completes the Gene Frequency Inverse Cell Frequency data transformation on a list of data. This step can be done to
create and use GFICF data for further downstream data analysis and computations. For more information on the GFICF calculation
please see https://github.com/dibbelab/gficf
}
