% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_harmony}
\alias{run_harmony}
\title{Run Harmony}
\usage{
run_harmony(dataList, batch_name = "ID", seed = 1)
}
\arguments{
\item{dataList}{A data list of data sets to integrate using the harmony protocol}
}
\value{
data.combined A data list of the combined data from the harmony protocol
}
\description{
This function runs the data integration protocol detailed in the Seurat ## FIND VING. AND LINK . This function completes the entire
protocol on either gficf or log transformed data.
}
