% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{run_pca}
\alias{run_pca}
\title{Run Principle Component Analysis}
\usage{
run_pca(dataList, numComponents = 30)
}
\arguments{
\item{dataList}{A data list of data sets}
}
\value{
A data list with PCA completed on the features
}
\description{
This functions runs the Seurat RunPCA function on a list of data sets. This functions assumes that genes are in rows and
cells are in columns. The default parameters are used, and verbose is set to false
}
