% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_sctransform}
\alias{run_sctransform}
\title{Run sctransform}
\usage{
run_sctransform(dataList, seed = 1)
}
\arguments{
\item{dataList}{A data list of data sets to integrate using the sctransform protocol}
}
\value{
data.combined A data list of the combined data from the sctransform protocol
}
\description{
This function runs the data integration protocol detailed in the Seurat "Using sctransform in Seurtat"
found at https://satijalab.org/seurat/articles/sctransform_vignette.html . This function completes the entire
protocol on either gficf or log transformed data.
}
