% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_sctransform}
\alias{run_sctransform}
\title{Run sctransform}
\usage{
run_sctransform(dataList, numFeatures = 2000)
}
\arguments{
\item{dataList}{A list of data sets to integrate using the sctransform protocol}

\item{numFeatures}{The number of Features to use, defaults to 2000}
}
\value{
A data list of the combined data from the sctransform protocol
}
\description{
This function runs the data integration protocol detailed in the Seurat "Using sctransform in Seurtat"
found at https://satijalab.org/seurat/articles/sctransform_vignette.html . This function completes the sctransform process on
the datasets. Use run_sctransform_workflow() for the entire workflow process
}
