% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{run_silhouette}
\alias{run_silhouette}
\title{Run Silhouette}
\usage{
run_silhouette(dataList, reduction_choosen = "pca", method = "kmeans")
}
\arguments{
\item{dataList}{a data list of the data with clustering assignments}

\item{reduction_choosen}{The dimensionality reduction type of the data to find the calculated scores}

\item{method}{The clustering method used}
}
\description{
This function calculates the average silhouette width of the clustering assignments made on the datasets provided
}
