% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{run_tsne}
\alias{run_tsne}
\title{Run tSNE}
\usage{
run_tsne(dataList, reductionChoosen = "pca", numDimensions = 30)
}
\arguments{
\item{dataList}{A data list of data sets}
}
\value{
A data list with tSNE completed on the features
}
\description{
This functions runs the Seurat RunTSNE function on a list of data sets. This functions assumes that genes are in rows and
cells are in columns. The default reduction is set to "pca" with the first 30 dimensions being accepted
}
