% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{run_umap}
\alias{run_umap}
\title{Run Uniform Manifold Aproximation and Projection}
\usage{
run_umap(data.list, reduction_choosen = "pca")
}
\arguments{
\item{data.list}{A data list of data sets}
}
\value{
A data list with UMAP completed on the features
}
\description{
This functions runs the Seurat RunUMAP function on a list of data sets. This functions assumes that genes are in rows and
cells are in columns. The default reduction is set to "pca" with the first 30 dimensions being accepted
}
