% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{select_hvg}
\alias{select_hvg}
\title{Select Highly Variable Genes}
\usage{
select_hvg(dataList, numFeatures = 2000)
}
\arguments{
\item{dataList}{A list of data sets that you wouuld like the variable genes of}
}
\value{
A data list of that includes the variable genes of the original data list
}
\description{
This function uses the FindVariablesFeatures of Seurat to find the Highly variable genes of a data set. This function
assumes that genes are in rows and cells are in columns. The "vst" selection method is used with 2000 nfeatures.
}
